+++ CHANGELOG.md
@@ -10,6 +10,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - [server] Database initial migration on mysql 5.7
 - [server] Doesn't log request content on error middleware if the route is `User/Create`
 
+### Changed
+- [server] Big refactor on spec pkg
+
 ## [0.24.1] - 2018-07-11
 ### Fixed
 - [server] Channel leak on watch deploy rolling update and service creation