+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [NEXT_RELEASE]
+## [0.25.1] - 2018-08-02
 ### Fixed
 - [server] Ingress setting for out of cluster configs
 - [server] `enable-ssl` command resetting the protocol to tcp